+++ Control/DeepSeq.hs
@@ -430,28 +430,31 @@ instance NFData1 Ratio where
   liftRnf r x = r (numerator x) `seq` r (denominator x)
 
 -- | @since 1.4.3.0
-instance (NFData1 f, NFData1 g) => NFData1(Compose f g) where liftRnf r = liftRnf (liftRnf r) . getCompose
+instance (NFData1 f, NFData1 g) => NFData1 (Compose f g) where
+  liftRnf r = liftRnf (liftRnf r) . getCompose
 
 -- | @since 1.4.3.0
-instance (NFData1 f, NFData1 g, NFData a) => NFData (Compose f g a) where rnf = rnf1
+instance (NFData1 f, NFData1 g, NFData a) => NFData (Compose f g a) where
+  rnf = rnf1
 
 -- | @since 1.4.3.0
-instance (NFData1 f, NFData1 g) => NFData1(Functor.Sum f g) where
+instance (NFData1 f, NFData1 g) => NFData1 (Functor.Sum f g) where
   liftRnf rnf0 (Functor.InL l) = liftRnf rnf0 l
   liftRnf rnf0 (Functor.InR r) = liftRnf rnf0 r
 
 -- | @since 1.4.3.0
-instance (NFData1 f, NFData1 g, NFData a) => NFData(Functor.Sum f g a) where rnf = rnf1
+instance (NFData1 f, NFData1 g, NFData a) => NFData (Functor.Sum f g a) where
+  rnf = rnf1
 
 -- | @since 1.4.3.0
-instance (NFData1 f, NFData1 g) => NFData1(Functor.Product f g) where
+instance (NFData1 f, NFData1 g) => NFData1 (Functor.Product f g) where
   liftRnf rnf0 (Functor.Pair f g) = liftRnf rnf0 f `seq` liftRnf rnf0 g
 
 -- | @since 1.4.3.0
-instance (NFData1 f, NFData1 g, NFData a) => NFData(Functor.Product f g a) where rnf = rnf1
+instance (NFData1 f, NFData1 g, NFData a) => NFData (Functor.Product f g a) where
+  rnf = rnf1
 
 instance NFData a => NFData (Ratio a) where
-
 #else
 instance (Integral a, NFData a) => NFData (Ratio a) where
 #endif